rotate_items_left test 8

+++ cse351/lab1b/lab1b/aisle_manager.c
@@ -285,8 +285,8 @@ void rotate_items_left(unsigned long *aisle, int index, int n)
 {
   // TODO: implement this method
   int rotation = n % 10;
-  unsigned int spaces = get_spaces(aisle, index);
-  unsigned short new_spaces = (spaces << rotation) | (spaces >> NUM_SPACES);
+  unsigned long spaces = get_spaces(aisle, index);
+  unsigned long new_spaces = (spaces << rotation) | (spaces >> NUM_SPACES);
   print_binary_short(spaces);
   print_binary_short(new_spaces);
   set_spaces(aisle, index, new_spaces);
